package restructuring and context based http handlers
